feat: resolve circular deps
JIRA: TNT-1700
