+++ libs/api-client-bear/src/caching/index.ts
@@ -1,0 +1,3 @@
+// (C) 2023 GoodData Corporation
+export * from "./types.js";
+export { withCaching, RecommendedCachingConfiguration } from "./cachingClient.js";
