+++ libs/api-client-bear/src/gooddata-node.ts
@@ -4,11 +4,7 @@ import fetchCookie from "fetch-cookie";
 import nodeFetch from "node-fetch";
 
 import { factory, SDK } from "./gooddata.js";
-export {
-    CachingConfiguration,
-    withCaching,
-    RecommendedCachingConfiguration,
-} from "./caching/cachingClient.js";
+export * from "./caching/index.js";
 
 // Calling getFactoryNode always reinitializes fetch-cookie. This way, when factory
 // initialization happens between requests, new instance of fetch-cookie and uderlying
